Convert Data to compoundIcrData Class

Converts a list object or several data.frames of FTICR data to an object of the class 'compoundIcrData'. Objects of the class 'compoundIcrData' are lists with three obligatory components e_data, f_data, and e_meta.

as.compoundIcrData(e_data, f_data, e_meta, edata_cname, fdata_cname,
  mass_cname, compound_cname, ...)

Arguments

e_data
a \(p \times n + 1\) data.frame of expression data, where \(p\) is the number of observed compounds and \(n\) is the number of samples. Each row corresponds to data for each peak. One column specifying a unique identifier for each peak/mass (row) must be present.

f_data
a data.frame with \(n\) rows. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.

e_meta
a data.frame with \(p\) rows. Each row corresponds to a compound with one column giving a unique peak/identifier (must be named the same as the column in e_data) and other columns giving meta information. At a minimum a column giving the mass of each peak and a column giving molecular formulae or columns giving elemental counts must be present.

edata_cname
character string specifying the name of the column containing a unique identifier for each peak/mass in e_data and e_meta.

fdata_cname
character string specifying the name of the column containing the sample identifiers in f_data.

mass_cname
character string specifying the name of the column containing the peak/mass identifiers in e_meta. Note: this is often the same as edata_cname for cases where mass is used as a unique identifier.

compound_cname
character string specifying the name of the column containing the compound identifier in e_meta. This is a compound identifier related to a database (e.g. Kegg, MetaCyc)

...
further arguments (see as.peakIcrData)

Details

as.compoundIcrData constructs a compoundIcrData object which is an icrData object where the rows of e_data correspond to compounds.