Default cognostics for Kendrick plots in Trelliscope

The kendrickCognostics function provides a set of default cognostics to be used with Kendrick plots in Trelliscope. The kendrickCognostics function accepts the boundary set used for Van Krevelen class calculations and (for comparisonSummary objects only) the name of the column to use for identifying which peaks are observed in which group. It returns a function that may be applied to each icrData object, as is appropriate for use with the makeDisplay function. See Examples section for use.

kendrickCognostics(vkBoundarySet = "bs1", uniquenessColName = NA)

Arguments

vkBoundarySet
Van Krevelen boundary set to use for calculating class proportions

uniquenessColName
if icrData is a group comparison summary object, what is the name of the column that specifies uniqueness to a group? If only one uniqueness function has been applied this is unnecessary. (See summarizeComparisons.)

Value

a function that may be applied to objects of type peakIcrData, groupSummary, and comparisonSummary

Examples

# NOT RUN {
library(fticRanalysis)
library(trelliscope)

vdbDir <- vdbConn(file.path(tempdir(), "trell_test"), autoYes = TRUE)
data('peakIcrProcessed')

## Kendrick plot for each sample
sampleDdo <- divideBySample(peakIcrProcessed)
panelFn1 <- panelFunctionGenerator("kendrickPlot", vkBoundarySet="bs2", title="Test")

# Note: make sure the same vkBoundarySet value is provided to the panel and cognostics functions
makeDisplay(sampleDdo,
         panelFn=panelFn1,
         cogFn=kendrickCognostics(vkBoundarySet="bs2"),
         name = "Kendrick_plots_per_sample")

## Kendrick plots for group summaries
groupDdo <- divideByGroup(peakIcrProcessed)
groupSummaryDdo <- summarizeGroups(groupDdo, summary_functions = c("prop_present", "n_present"))

panelFn2 <- panelFunctionGenerator("kendrickPlot", colorCName=expr(paste0(getSplitVar(v, "Group"), "_n_present")),
                                 legendTitle="Number<br>Present")

makeDisplay(grpCompSummaryDdo,
         panelFn=panelFn2,
         cogFn=kendrickCognostics(),
         name = "Kendrick_plots_for_group_summaries")

view()
# }