Processed FTICR Data Object of Class peakIcrData

An S3 object of class peakIcrData

Format

A peakIcrData object (see as.peakIcrData for details)

e_data

a \(p \times n + 1\) data.frame of expression data, where \(p\) is the number of peaks observed and \(n\) is the number of samples. Each row corresponds to data for each peak/mass

f_data

a data.frame with \(n\) s. Each row corresponds to a sample with one column giving the unique sample identifiers found in e_data column names and other columns providing qualitative and/or quantitative traits of each sample.

e_meta

a \(p \times 10\) data.frame of meta information for each peak/mass.

Details

Created by running the following commands:

peakIcrProcessed <- group_designation(peakIcrData, main_effects=c("Location", "Crop.Flora"))
peakIcrProcessed <- compound_calcs(peakIcrProcessed)
peakIcrProcessed <- applyFilt(mass_filter(peakIcrProcessed), peakIcrProcessed, min_mass=200, max_mass=900)
peakIcrProcessed <- applyFilt(molecule_filter(peakIcrProcessed), peakIcrProcessed, min=2)